Organize CLI for subcmd
