+++ .github/workflows/ecosystem-main_repos_fetch.yml
@@ -29,7 +29,7 @@ jobs:
         pip install -r requirements-dev.txt
 
     - name: Fetch test results
-      run: python manager.py fetch_and_update_main_tests_results
+      run: python manager.py tests fetch_and_update_main_tests_results
 
     - name: Create PR for batch checks
       id: cpr
