+++ .github/workflows/ecosystem-recompile.yml
@@ -37,7 +37,7 @@ jobs:
  
     - name: Recompile and push
       run: |
-        python -m manager recompile
+        python -m manager members recompile
         git config user.name github-actions
         git config user.email github-actions@github.com
