+++ docs/project_overview.md
@@ -31,12 +31,12 @@ Entrypoint is ``manager.py`` file in the root of repository.
 
 Example of commands:
 ```shell
-python manager.py python_dev_tests https://github.com/IceKhan13/demo-implementation --python_version=py39
-python manager.py python_stable_tests https://github.com/IceKhan13/demo-implementation --python_version=py39
+python manager.py tests python_dev_tests https://github.com/IceKhan13/demo-implementation --python_version=py39
+python manager.py tests python_stable_tests https://github.com/IceKhan13/demo-implementation --python_version=py39
 ```
 or in general
 ```shell
-python manager.py <NAME_OF_FUNCTION_IN_MANAGER_FILE> <POSITIONAL_ARGUMENT> [FLAGS]
+python manager.py <CMD> <NAME_OF_FUNCTION_IN_MANAGER_FILE> <POSITIONAL_ARGUMENT> [FLAGS]
 ```
 
 ### Adding project to the ecosystem
@@ -119,13 +119,13 @@ file. This file should be compiled automatically by an action on pushing to
 `main`, but you can also compile it locally (e.g. for testing) using
 
 ```sh
-python -m manager recompile
+python -m manager members recompile
 ```
 
 You shouldn't edit `members.json` manually.
 
 If you somehow get a merge conflict in `members.json`, don't try to manually
-resolve the conflict. Instead, merge the branch, then run `python -m manager
+resolve the conflict. Instead, merge the branch, then run `python -m manager members
 recompile` and add the file to resolve the conflict.
 
 ### Tests
@@ -136,21 +136,21 @@ There are 3 type of tests for project: `STANDARD`, `DEV` and `STABLE`.
 
 CLI command:
 ```shell
-python manager.py python_standard_tests <REPO_URL> --run_name=<NAME_OF_RUN> --python_version=py39 --tier=<TIER>
+python manager.py tests python_standard_tests <REPO_URL> --run_name=<NAME_OF_RUN> --python_version=py39 --tier=<TIER>
 ```
 
 `DEV` - runs tests with default requirements for project + dev version of qiskit-terra installed
 
 CLI command:
 ```shell
-python manager.py python_dev_tests <REPO_URL> --run_name=<NAME_OF_RUN> --python_version=py39 --tier=<TIER>
+python manager.py tests python_dev_tests <REPO_URL> --run_name=<NAME_OF_RUN> --python_version=py39 --tier=<TIER>
 ```
 
 `STABLE` - runs tests with default requirements for project + latest stable version of qiskit-terra installed
 
 CLI command:
 ```shell
-python manager.py python_stable_tests <REPO_URL> --run_name=<NAME_OF_RUN> --python_version=py39 --tier=<TIER>
+python manager.py tests python_stable_tests <REPO_URL> --run_name=<NAME_OF_RUN> --python_version=py39 --tier=<TIER>
 ```
 
 You can see full setup on test running in [GitHub Action](https://github.com/qiskit-community/ecosystem/blob/main/.github/actions/run-tests/action.yml)
