+++ ecosystem/__init__.py
@@ -1,2 +1,1 @@
 """Ecosystem main module."""
-from .manager import Manager
