+++ ecosystem/cli/__init__.py
@@ -1,0 +1,5 @@
+"""CLI."""
+from .members import CliMembers
+from .website import CliWebsite
+from .ci import CliCI
+from .tests import CliTests
