+++ tox.ini
@@ -37,7 +37,7 @@ commands = black {posargs} ecosystem tests --check
 allowlist_externals = bash
 basepython = python3
 commands =
-  bash -ec "python manager.py build_website > website/index.html"
+  bash -ec "python manager.py website build_website > website/index.html"
 
 [testenv:ci]
 commands =
